business/1_b_DocMiner/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.15
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
langdetect==1.0.9
pydantic==2.5.0
pydantic-settings==2.1.0c
python-dotenv==1.0.0

